requirements.txt
fastapi
instructor
openai
pydantic
python-dotenv
SQLAlchemy
uvicorn
